learn-go-with-tests/5-structs/struct.go
package structs

type Shape interface {
	Area() float64
}

type Rectangle struct {
	Width  float64
	Height float64
}

func (r Rectangle) Area() float64 {
	return r.Width * r.Height
}

type Circle struct {
	Radius float64
}

func (r Circle) Area() float64 {
	return r.Radius * r.Radius * 3.141592653589793
}

func Perimeter(rectangle Rectangle) (p float64) {
	p = rectangle.Width*2 + rectangle.Height*2
	return
}

type Triangle struct {
	Base   float64
	Height float64
}

func (t Triangle) Area() float64 {
	return t.Base * t.Height / 2
}
